# Conflicts:
#	.gitignore
#	Dockerfile
#	README.md
#	cbpi/__init__.py
#	cbpi/api/actor.py
#	cbpi/api/base.py
#	cbpi/api/dataclasses.py
#	cbpi/api/fermenter_logic.py
#	cbpi/api/kettle_logic.py
#	cbpi/api/sensor.py
#	cbpi/api/timer.py
#	cbpi/cli.py
#	cbpi/configFolder.py
#	cbpi/controller/actor_controller.py
#	cbpi/controller/basic_controller2.py
#	cbpi/controller/config_controller.py
#	cbpi/controller/dashboard_controller.py
#	cbpi/controller/fermentation_controller.py
#	cbpi/controller/fermenter_recipe_controller.py
#	cbpi/controller/kettle_controller.py
#	cbpi/controller/log_file_controller.py
#	cbpi/controller/notification_controller.py
#	cbpi/controller/plugin_controller.py
#	cbpi/controller/recipe_controller.py
#	cbpi/controller/satellite_controller.py
#	cbpi/controller/sensor_controller.py
#	cbpi/controller/step_controller.py
#	cbpi/controller/system_controller.py
#	cbpi/controller/upload_controller.py
#	cbpi/craftbeerpi.py
#	cbpi/extension/ConfigUpdate/__init__.py
#	cbpi/extension/FermenterHysteresis/__init__.py
#	cbpi/extension/dummyactor/__init__.py
#	cbpi/extension/gpioactor/__init__.py
#	cbpi/extension/mashstep/__init__.py
#	cbpi/extension/mqtt_util/__init__.py
#	cbpi/http_endpoints/http_actor.py
#	cbpi/http_endpoints/http_config.py
#	cbpi/http_endpoints/http_dashboard.py
#	cbpi/http_endpoints/http_fermentation.py
#	cbpi/http_endpoints/http_fermenterrecipe.py
#	cbpi/http_endpoints/http_kettle.py
#	cbpi/http_endpoints/http_log.py
#	cbpi/http_endpoints/http_notification.py
#	cbpi/http_endpoints/http_plugin.py
#	cbpi/http_endpoints/http_recipe.py
#	cbpi/http_endpoints/http_sensor.py
#	cbpi/http_endpoints/http_step.py
#	cbpi/http_endpoints/http_system.py
#	cbpi/http_endpoints/http_upload.py
#	cbpi/job/_job.py
#	cbpi/job/_scheduler.py
#	cbpi/utils/encoder.py
#	cbpi/websocket.py
#	requirements.txt
#	setup.py

+++ cbpi/http_endpoints/http_system.py
@@ -1,34 +1,22 @@
-"""http_system.py - REST-API Endpunkte fuer Systemverwaltung
-
-Routen:
-    GET    /system/              - Gesamter Systemstatus (Aktoren, Sensoren, Kessel, Config)
-    GET    /system/logs           - Logdateien auflisten
-    GET    /system/events         - Registrierte Events auflisten
-    GET    /system/jobs           - Laufende Hintergrund-Jobs
-    GET    /system/backup         - Konfigurations-Backup herunterladen (ZIP)
-    GET    /system/log/{logtime}/ - Komprimiertes Log-Paket herunterladen
-    POST   /system/restore        - Backup wiederherstellen
-    POST   /system/restart        - System neu starten
-    POST   /system/shutdown       - System herunterfahren
-"""
-
-import re
-from aiohttp import web
-from aiohttp import streamer
-from cbpi.job.aiohttp import get_scheduler_from_app
 import logging
-from cbpi.api import request_mapping
-from cbpi.utils import json_dumps
-from cbpi import __version__, __codename__
-import pathlib
 import os
+import pathlib
+import re
+import psutil
+
+from aiohttp import streamer, web
+from cbpi import __codename__, __version__
+from cbpi.api import request_mapping
 from cbpi.controller.system_controller import SystemController
+from cbpi.job.aiohttp import get_scheduler_from_app
+from cbpi.utils import json_dumps
+
 
 class SystemHttpEndpoints:
 
-    def __init__(self,cbpi):
+    def __init__(self, cbpi):
         self.cbpi = cbpi
-        self.controller : SystemController = cbpi.system
+        self.controller: SystemController = cbpi.system
         self.cbpi.register(self, url_prefix="/system")
 
     @request_mapping("/", method="GET", auth_required=False)
@@ -42,30 +30,55 @@ class SystemHttpEndpoints:
             "200":
                 description: successful operation
         """
-        return web.json_response(data=dict(
-            actor=self.cbpi.actor.get_state(),
-            fermenter=self.cbpi.fermenter.get_state(),
-            sensor=self.cbpi.sensor.get_state(),
-            kettle=self.cbpi.kettle.get_state(),
-            step=self.cbpi.step.get_state(),
-            fermentersteps=self.cbpi.fermenter.get_fermenter_steps(),
-            config=self.cbpi.config.get_state(),
-            version=__version__,
-            codename=__codename__)
-            , dumps=json_dumps)
+        plugin_list = await self.cbpi.plugin.load_plugin_list("cbpi4gui")
+        try:
+            version = plugin_list[0].get("Version", "not detected")
+        except:
+            version = "not detected"
+
+        spindle = self.cbpi.config.get("spindledata", "No")
+        if spindle == "Yes":
+            spindledata = True
+        else:
+            spindledata = False
+        
+        mem = psutil.virtual_memory()
+
+        return web.json_response(
+            data=dict(
+                actor=self.cbpi.actor.get_state(),
+                fermenter=self.cbpi.fermenter.get_state(),
+                sensor=self.cbpi.sensor.get_state(),
+                kettle=self.cbpi.kettle.get_state(),
+                step=self.cbpi.step.get_state(),
+                fermentersteps=self.cbpi.fermenter.get_fermenter_steps(),
+                config=self.cbpi.config.get_state(),
+                notifications=self.cbpi.notification.get_state(),
+                bf_recipes=await self.cbpi.upload.get_brewfather_recipes(0),
+                version=__version__,
+                spindledata=spindledata,
+                guiversion=version,
+                codename=__codename__,
+                system=dict(
+                    totalmem=round((int(mem.total) / (1024 * 1024)), 1),
+                    availablemem=round((int(mem.available) / (1024 * 1024)), 1),
+                    percentmem=round(float(mem.percent), 1),
+                )),
+            dumps=json_dumps,
+        )
 
     @request_mapping(path="/logs", auth_required=False)
     async def http_get_log(self, request):
         result = []
         file_pattern = re.compile("^(\w+.).log(.?\d*)")
-        for filename in sorted(os.listdir("./logs"), reverse=True):  #
+        for filename in sorted(os.listdir(self.cbpi.logsFolderPath), reverse=True):
             if file_pattern.match(filename):
                 result.append(filename)
         return web.json_response(result)
 
     @request_mapping(path="/logs/{name}", method="DELETE", auth_required=False)
     async def delete_log(self, request):
-        log_name = request.match_info['name']
+        log_name = request.match_info["name"]
         self.cbpi.log.delete_log(log_name)
 
     @request_mapping(path="/logs", method="DELETE", auth_required=False)
@@ -73,40 +86,36 @@ class SystemHttpEndpoints:
         self.cbpi.log.delete_logs()
         return web.Response(status=204)
 
-    @request_mapping("/events", method="GET", name="get_all_events", auth_required=False)
-    def get_all_events(self, request):
+    @request_mapping(
+        "/events", method="GET", name="get_all_events", auth_required=False
+    )
+    async def get_all_events(self, request):
         """
         ---
-        description: Get list of all registered events
+        description: Get all events in the event bus
         tags:
         - System
         responses:
             "200":
                 description: successful operation
         """
-        return web.json_response(data=self.cbpi.bus.dump())
+        data = self.cbpi.bus.dump()
+        return web.json_response(data=data)
 
     @request_mapping("/jobs", method="GET", name="get_jobs", auth_required=False)
-    def get_all_jobs(self, request):
-        """
-        ---
-        description: Get all running Jobs
-        tags:
-        - System
-        responses:
-            "200":
-                description: successful operation
-        """
+    async def get_all_jobs(self, request):
         scheduler = get_scheduler_from_app(self.cbpi.app)
         result = []
         for j in scheduler:
             try:
                 result.append(dict(name=j.name, type=j.type, time=j.start_time))
-            except (AttributeError, TypeError):
+            except:
                 pass
         return web.json_response(data=result)
 
-    @request_mapping("/restart", method="POST", name="RestartServer", auth_required=False)
+    @request_mapping(
+        "/restart", method="POST", name="RestartServer", auth_required=False
+    )
     async def restart(self, request):
         """
         ---
@@ -120,7 +129,9 @@ class SystemHttpEndpoints:
         await self.controller.restart()
         return web.Response(text="RESTART")
 
-    @request_mapping("/shutdown", method="POST", name="ShutdownSerer", auth_required=False)
+    @request_mapping(
+        "/shutdown", method="POST", name="ShutdownSerer", auth_required=False
+    )
     async def shutdown(self, request):
         """
         ---
@@ -138,7 +149,7 @@ class SystemHttpEndpoints:
     async def backup(self, request):
         """
         ---
-        description: Zip and download Config Folder 
+        description: Zip and download Config Folder
         tags:
         - System
         responses:
@@ -147,64 +158,91 @@ class SystemHttpEndpoints:
                 content:  # Response body
                 application/zip:  # Media type
         """
-        await self.controller.backupConfig()
-        filename = "cbpi4_config.zip"
+        filename = await self.controller.backupConfig()
+        # filename = "cbpi4_config.zip"
         file_name = pathlib.Path(os.path.join(".", filename))
 
         response = web.StreamResponse(
             status=200,
-            reason='OK',
-            headers={'Content-Type': 'application/zip'},
+            reason="OK",
+            headers={"Content-Type": "application/zip"},
         )
         await response.prepare(request)
-        with open(file_name, 'rb') as file:
+        with open(file_name, "rb") as file:
             for line in file.readlines():
                 await response.write(line)
 
         await response.write_eof()
         return response
 
-
-    @request_mapping("/log/{logtime}/", method="GET", name="BackupConfig", auth_required=False)
+    @request_mapping(
+        "/log/{logtime}/", method="GET", name="BackupConfig", auth_required=False
+    )
     async def downloadlog(self, request):
         """
         ---
-        description: Zip and download craftbeerpi.service log 
+        description: Zip and download craftbeerpi.service log
         tags:
         - System
+        parameters:
+        - name: "logtime"
+          in: "path"
+          description: "Logtime in hours"
+          required: true
+          type: "integer"
+          format: "int64"
         responses:
             "200":
                 description: successful operation
                 content:  # Response body
                 application/zip:  # Media type
         """
-        logtime = request.match_info['logtime']
-        await self.controller.downloadlog(logtime)
-        filename = "cbpi4_log.zip"
-        file_name = pathlib.Path(os.path.join(".", filename))
+        checklogtime = False
+        logtime = request.match_info["logtime"]
 
-        response = web.StreamResponse(
-            status=200,
-            reason='OK',
-            headers={'Content-Type': 'application/zip'},
-        )
-        await response.prepare(request)
-        with open(file_name, 'rb') as file:
-            for line in file.readlines():
-                await response.write(line)
+        try:
+            test = int(logtime)
+            checklogtime = True
+        except:
+            if logtime == "b":
+                checklogtime = True
 
-        await response.write_eof()
-        os.remove(file_name)
-        return response
+        if checklogtime:
+            await self.controller.downloadlog(logtime)
+            filename = "cbpi4_log.zip"
+            file_name = pathlib.Path(os.path.join(".", filename))
 
+            response = web.StreamResponse(
+                status=200,
+                reason="OK",
+                headers={"Content-Type": "application/zip"},
+            )
+            await response.prepare(request)
+            with open(file_name, "rb") as file:
+                for line in file.readlines():
+                    await response.write(line)
 
-    @request_mapping("/restore", method="POST", name="RestoreConfig", auth_required=False)
+            await response.write_eof()
+            os.remove(file_name)
+            return response
+        else:
+            return web.Response(status=400, text='Need integer or "b" for logtime.')
+
+    @request_mapping(
+        "/restore", method="POST", name="RestoreConfig", auth_required=False
+    )
     async def restore(self, request):
         """
         ---
         description: Restore Config
         tags:
         - System
+        parameters:
+        - name: "File"
+          in: "formData"
+          description: "Config data to restore"
+          required: true
+          type: "file"
         responses:
             "200":
                 description: successful operation
@@ -215,7 +253,9 @@ class SystemHttpEndpoints:
         await self.controller.restoreConfig(data)
         return web.Response(status=200)
 
-    @request_mapping("/systeminfo", method="GET", name="SystemInfo", auth_required=False)
+    @request_mapping(
+        "/systeminfo", method="GET", name="SystemInfo", auth_required=False
+    )
     async def systeminfo(self, request):
         """
         ---
@@ -233,9 +273,15 @@ class SystemHttpEndpoints:
     async def uploadSVG(self, request):
         """
         ---
-        description: Upload SVG file to widgets folder 
+        description: Upload SVG file to widgets folder
         tags:
         - System
+        parameters:
+        - name: "File"
+          in: "formData"
+          description: "SVG file to upload"
+          required: true
+          type: "file"
         responses:
             "200":
                 description: successful operation
